LayoutTests/platform/gtk/fullscreen/exit-full-screen-video-crash-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: video0.webkitSetPresentationMode is not a function. (In 'video0.webkitSetPresentationMode('inline')', 'video0.webkitSetPresentationMode' is undefined)
